feat: support connecting to emulator over unix domain socket
Getting the emulator itself listening on a unix domain socket can be
accomplished by some proxy/socat, some LD_PRELOAD magic (ip2unix) or
using a patch as proposed in
https://github.com/googleapis/google-cloud-go/pull/9665

```
$ ./emulator --address $PWD/emulator.sock &
Cloud Bigtable emulator running on /home/flokli/dev/google-cloud-go/bigtable/cmd/emulator/emulator.sock
$ export BIGTABLE_EMULATOR_HOST=unix://$PWD/emulator.sock
$ cbt -instance instance-1 -project project-1 createtable table-1
2024/03/29 11:02:04 -creds flag unset, will use gcloud credential
$ cbt -instance instance-1 -project project-1 ls
2024/03/29 11:02:15 -creds flag unset, will use gcloud credential
table-1
```

As can be seen through the `cbt` invocations, the golang SDK already
supports connecting to an emulator listening at a unix domain socket.

This commit adds support to connect to a unix domain socket to
bigtable-rs too.

`create_client` now accepts a channel, rather than a list of endpoints.
The constructors call `Channel::balance_list(endpoints.into_iter())`
themselves, in case of a normal bigtable connection.`

In case of a emulator connection, we produce a channel directly
connecting to the emulator, either via HTTP, or via Unix Socket, instead
of using a single endpoint balancer. We use the lazy counterpart to keep
behaviour (and function signatures) consistent with the previous
`Channel::balance_list`.

In the uds case, we use a custom connector, similar to the example in
https://github.com/hyperium/tonic/pull/184/files#diff-f27114adeedf7b42e8656c8a86205685a54bae7a7929b895ab62516bdf9ff252.

Additionally, `new_with_emulator` is made public, and a docstring is
added. It allows connecting to the emulator explicitly, in cases where
setting BIGTABLE_EMULATOR_HOST is inconvenient, for example when
connecting to a (new, separate) emulator during unit tests.