feat: Implement wrapper for execute_query call in BigTable struct
This exposes the execute_query call from the generated BigtableClient service.
It also adds the neccessary routing headers (AIP-4222) that are specified in
the service declaration in the source Protobuf. Without those, calling the
execute_query always fails with 400 Unavailable. Unfortunately neither Prost
or Tonic support the google.api.routing annotation for gRPC Services, so we must
add it manually like this. Technically, the option is present for all the
Bigtable service calls, but I suspect for all the other calls the server can
fallback to obtaining all the needed information from the Request message.
In case of ExecuteQueryRequest, the table name is only specified inside the
query string. This way I was able to successfully execute a basic BTQL query.
I haven't tested more complex use-cases like non-empty app_profile_id.